tag: changing ClassLoader to ClassPathResource

+++ src/main/java/com/shopspace/storage/util/FirebaseStorageUtil.java
@@ -123,6 +123,8 @@ public class FirebaseStorageUtil {
         try {
             Blob exists = this.getBlobFile(name, folder);
 
+            if (exists == null) return false;
+
             return exists.delete();
         } catch (Exception e) {
             e.printStackTrace();
